[REF] web_tour: tour_service to plugin
Converts TourService into an owl3 Plugin (TourPlugin), keeping the
legacy tour_service registration as a compatibility bridge. Extracts
the tour recorder overlay/command-provider logic into their own
TourRecorderPlugin and the debug-menu OnboardingItem widget.
