+++ addons/web_tour/__manifest__.py
@@ -63,10 +63,13 @@ Odoo Web tours.
             'web_tour/static/src/tour_pointer/tour_pointer.xml',
             'web_tour/static/src/tour_helpers/tour_helpers.js',
             'web_tour/static/src/tour_state.js',
-            'web_tour/static/src/tour_service.js',
+            'web_tour/static/src/tour_plugin.js',
             'web_tour/static/src/tour_recorder/tour_recorder_state.js',
+            'web_tour/static/src/tour_recorder/tour_recorder_plugin.js',
+            'web_tour/static/src/tour_recorder/tour_recorder_command_provider.js',
             'web_tour/static/src/tour_utils.js',
-            'web_tour/static/src/onboarding_item.xml',
+            'web_tour/static/src/widgets/onboarding_item.js',
+            'web_tour/static/src/widgets/onboarding_item.xml',
         ],
     },
     'auto_install': True,
