+++ addons/web_tour/static/src/widgets/onboarding_item.js
@@ -1,0 +1,12 @@
+import { Component, t, useProps } from "@odoo/owl";
+import { DropdownItem } from "@web/core/dropdown/dropdown_item";
+
+export class OnboardingItem extends Component {
+    static components = { DropdownItem };
+    static template = "web_tour.OnboardingItem";
+    props = useProps({
+        toursEnabled: t.boolean(),
+        toggleItem: t.function(),
+    });
+    setup() {}
+}
