+++ addons/web_tour/static/src/widgets/onboarding_item.xml
